[CT-17] refactor: common 패키지 이름변경, 힐트모듈 이름변경, ktlintFormat
