+++ data/src/main/java/com/yessorae/data/source/network/ChartNetworkDataSource.kt
@@ -1,6 +1,6 @@
 package com.yessorae.data.source.network
 
-import com.yessorae.domain.entity.Chart
+import com.yessorae.data.source.network.polygon.model.chart.ChartDto
 import com.yessorae.domain.entity.tick.TickUnit
 
 interface ChartNetworkDataSource {
@@ -9,5 +9,5 @@ interface ChartNetworkDataSource {
         tickUnit: TickUnit,
         from: String,
         to: String
-    ): Chart
+    ): ChartDto
 }
